+++ dbms/src/Flash/Mpp/ExchangeReceiver.h
@@ -126,7 +126,7 @@ public:
     ExchangeReceiverResult nextResult(
         std::queue<Block> & block_queue,
         const Block & header,
-        UInt32 stream_id);
+        size_t stream_id);
 
     size_t getSourceNum() const { return source_num; }
     uint64_t getFineGrainedShuffleStreamCount() const { return fine_grained_shuffle_stream_count; }
